requirements.txt
streamlit==1.38
pandas
numpy
openai==1.40.0
langchain-openai==0.2.0
langchain-core==0.3.0
langchain-community
tiktoken
faiss-cpu
pyarrow
python-docx
openpyxl
python-pptx
pypdf
xlrd
cryptography
pycryptodome
beautifulsoup4==4.12.3
requests==2.31.0
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
tenacity